ocr update

+++ src/main/java/com/example/tomyongji/receipt/service/ReceiptService.java
@@ -1,5 +1,12 @@
 package com.example.tomyongji.receipt.service;
 
+import static com.example.tomyongji.validation.ErrorMsg.DUPLICATED_FLOW;
+import static com.example.tomyongji.validation.ErrorMsg.EMPTY_CONTENT;
+import static com.example.tomyongji.validation.ErrorMsg.EMPTY_MONEY;
+import static com.example.tomyongji.validation.ErrorMsg.NOT_FOUND_RECEIPT;
+import static com.example.tomyongji.validation.ErrorMsg.NOT_FOUND_STUDENT_CLUB;
+import static com.example.tomyongji.validation.ErrorMsg.NOT_FOUND_USER;
+
 import com.example.tomyongji.auth.entity.User;
 import com.example.tomyongji.auth.repository.UserRepository;
 import com.example.tomyongji.receipt.dto.ReceiptDto;
@@ -7,6 +14,8 @@ import com.example.tomyongji.receipt.entity.Receipt;
 import com.example.tomyongji.receipt.entity.StudentClub;
 import com.example.tomyongji.receipt.repository.ReceiptRepository;
 import com.example.tomyongji.receipt.repository.StudentClubRepository;
+import com.example.tomyongji.validation.CustomException;
+import java.util.ArrayList;
 import java.util.Calendar;
 import java.util.Date;
 import java.util.List;
@@ -32,7 +41,7 @@ public class ReceiptService {
     public ReceiptDto createReceipt(ReceiptDto receiptDto, Long id) {
         Optional<User> user = userRepository.findById(id);
         if (user.isEmpty()) {
-            throw new RuntimeException("유저를 찾을 수 없습니다.");
+            throw new CustomException(NOT_FOUND_USER, 400);
         }
         StudentClub studentClub = user.get().getStudentClub();
 
@@ -40,16 +49,16 @@ public class ReceiptService {
             receiptDto.setDate(new Date());
         }
         if (receiptDto.getContent() == null || receiptDto.getContent().trim().isEmpty()) {
-            throw new RuntimeException("내용을 적어주세요.");
+            throw new CustomException(EMPTY_CONTENT, 400);
         }
 
         // null 값을 0으로 처리
         if (receiptDto.getDeposit() == 0 && receiptDto.getWithdrawal() == 0) {
-            throw new RuntimeException("입출금 내역을 적어주세요.");
+            throw new CustomException(EMPTY_MONEY, 400);
         }
 
         if (receiptDto.getDeposit() != 0 && receiptDto.getWithdrawal() != 0) {
-            throw new RuntimeException("입금과 출금 둘 중 하나만 적어주세요.");
+            throw new CustomException(DUPLICATED_FLOW, 400);
         }
 
         Receipt receipt = convertToReceipt(receiptDto);
@@ -68,27 +77,35 @@ public class ReceiptService {
         return receiptDto;
     }
 
-    public List<Receipt> getAllReceipts() {
-        return receiptRepository.findAll();
+    public List<ReceiptDto> getAllReceipts() {
+        List<Receipt> receipts = receiptRepository.findAll();
+        return receiptDtoList(receipts);
     }
-    public List<Receipt> getReceiptsByClub(Long clubId) {
+    public List<ReceiptDto> getReceiptsByClub(Long clubId) {
 
         Optional<StudentClub> studentClub = studentClubRepository.findById(clubId);
         if (studentClub.isEmpty()) {
-            throw new RuntimeException("학생회를 찾을 수 없습니다.");
+            throw new CustomException(NOT_FOUND_STUDENT_CLUB, 400);
         }
 
-        return receiptRepository.findAllByStudentClub(studentClub.get());
+        List<Receipt> receipts = receiptRepository.findAllByStudentClub(studentClub.get());
+        return receiptDtoList(receipts);
     }
-    public Receipt getReceiptById(Long id) {
+    public ReceiptDto getReceiptById(Long id) {
         Optional<Receipt> receipt = receiptRepository.findById(id);
         if (receipt.isEmpty()) {
-            throw new RuntimeException("영수증을 찾을 수 없습니다.");
+            throw new CustomException(NOT_FOUND_RECEIPT, 400);
         }
-        return receipt.get();
+        return convertToReceiptDto(receipt.get());
     }
-    public void deleteReceipt(Long id) {
+    public ReceiptDto deleteReceipt(Long id) {
+        Optional<Receipt> receipt = receiptRepository.findById(id);
+        if (receipt.isEmpty()) {
+            throw new CustomException(NOT_FOUND_RECEIPT, 400);
+        }
+        ReceiptDto receiptDto = convertToReceiptDto(receipt.get());
         receiptRepository.deleteById(id);
+        return receiptDto;
     }
     public ReceiptDto updateReceipt(Long id, ReceiptDto receiptDto) {
         Optional<Receipt> optionalReceipt = receiptRepository.findById(id);
@@ -113,7 +130,7 @@ public class ReceiptService {
             receiptRepository.save(existingReceipt);
             return receiptDto;
         } else {
-            throw new RuntimeException("영수증을 찾을 수 없습니다.");
+            throw new CustomException(NOT_FOUND_RECEIPT, 400);
         }
     }
     public Receipt convertToReceipt(ReceiptDto receiptDto) {
@@ -123,6 +140,23 @@ public class ReceiptService {
         receipt.setDeposit(receiptDto.getDeposit());
         receipt.setWithdrawal(receiptDto.getWithdrawal());
         return receipt;
+    }
+
+    private ReceiptDto convertToReceiptDto(Receipt receipt) {
+        ReceiptDto receiptDto = new ReceiptDto();
+        receiptDto.setDate(receipt.getDate());
+        receiptDto.setContent(receipt.getContent());
+        receiptDto.setDeposit(receipt.getDeposit());
+        receiptDto.setWithdrawal(receipt.getWithdrawal());
+        return receiptDto;
+    }
+
+    private List<ReceiptDto> receiptDtoList(List<Receipt> receipts) {
+        List<ReceiptDto> receiptDtoList = new ArrayList<>();
+        for (Receipt receipt : receipts) {
+            receiptDtoList.add(convertToReceiptDto(receipt));
+        }
+        return receiptDtoList;
     }
